chore: fix order variants table, move background options to chip template
